Split connection tracking app into contrack app and contrack lib
